refactor: move object and array parsers into their own files
